отрисовывает заглушку пустой страницы избранного
